streamlit_app/fastapi_app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
openai==1.58.1
python-dotenv==1.0.1
python-multipart==0.0.16
requests==2.32.3
pillow==11.3.0
sqlalchemy==2.0.36
alembic==1.14.0
streamlit==1.41.1